requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.57.3

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Utilities
numpy>=1.24.0
scipy>=1.10.0

# Alert System
pygame>=2.5.0

# Jupyter (optional, untuk .ipynb)
jupyter>=1.0.0
ipykernel>=6.25.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0